refactor(chat): extract chat_manager/external_facade.rs (#292)
Moves ExternalSessionFacade and its blanket impl for
ExternalSessionService<D> into their own submodule.

Task 4 of the chat_manager.rs file-limit split plan.
